Fix only first display pattern being localized
Recall that display patterns are bindings that replace %{value1},
%{value2}, ... patterns in NSTextField strings.

The problem was caused by unbinding the display pattern and then
re-binding in order to update the binding with the localized string.
However, this unbinding unbound _all_ display patterns.

It turns out that unbinding is unecessary, so fixed by removing it.