Fix NSTokenField objects converted to strings
The problem was caused because NSTokenFields were handled as
NSTextField (their superclass), which meant that the setStringValue:
method was called.

However, this implicitly converts any custom (= non-string) token
objects to strings.

I noticed this because suddenly NSTokenFieldDelegate's
tokenField:representedObjectForEditingString: method was called when
launching the app in a localized language, but not called for the
English version.